requirements.txt
mysql-connector-python==8.0.21
streamlit==0.79.0
requests==2.25.1